+++ wow_world_messages/CHANGELOG.md
@@ -18,6 +18,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 * BREAKING: `Server/ClientOpcodeMessage` now boxes any types larger than 8 bytes in order to avoid breaking the stack.
 * BREAKING: Changed `BattlegroundBracket::Fourties` to `BattlegroundBracket::Forties`.
+* BREAKING: Changed `BuyResult::NotEnoughtMoney` to `BuyResult::NotEnoughMoney`.
 
 ### Removed